Added error checking for multiple subnode functionality. And added conditionality to the combine file so it can just be run without commenting out sections

+++ run/combine.py
@@ -72,39 +72,44 @@ else:
 
 
 
-subnodes=4
+subnodes=inputs.run['subnodes'] 
 multflg=inputs.run['multiple']
-
+stp1flg=1
 
 # If combining into temporary folders comment this all out if not using (for now) 
-folders=multflg/subnodes
-# Step 1
-# for l in range(1,subnodes):
-#     shutil.copy2(EXDIR+'/node_'+str(l+1)+'/data/zombie_1001.csv',EXDIR+'/node_'+str(l+1)+'_1/data/')
-#     cnt=2
-#     for i in range(int(folders*l),int((folders*l)+folders)):
-#         floc=EXDIR+'/node_'+str(i+1)+'/data/zombie_1'
-#         for j in range(2,inputs.zombs['ndet']+1):
-#             shutil.copy2(floc+"{:03d}".format(j)+".csv",EXDIR+'/node_'+str(l+1)+'_1/data/zombie_1'+"{:03d}".format(cnt)+".csv")
-#             cnt+=1
-
-
-# Step 2
-shutil.copy2(EXDIR+'/node_1_1/data/zombie_1001.csv',EXDIR+'/data')
-cnt=2
-for l in range(subnodes):
-    for j in range(2,int((folders*(inputs.zombs['ndet']-1))+2)):
-        shutil.copy2(EXDIR+'/node_'+str(l+1)+'_1/data/zombie_1'+"{:03d}".format(j)+".csv",EXDIR+'/data/zombie_1'+"{:03d}".format(cnt)+".csv")
-        cnt+=1
-
+if(subnodes>1):
+    folders=multflg/subnodes
+    if not os.path.exists("node_1_1"):
+        stp1flg=2
+        # Step 1
+        for l in range(1,subnodes):
+            os.mkdir(EXDIR+'/node_'+str(l+1)+'_1')
+            os.mkdir(EXDIR+'/node_'+str(l+1)+'_1/data')
+            shutil.copytree((EXDIR+'/integrals'),(EXDIR+'/node_'+str(l+1)+'_1/integrals'))
+            shutil.copy2(EXDIR+'/node_'+str(l+1)+'/data/zombie_1001.csv',EXDIR+'/node_'+str(l+1)+'_1/data/')
+            cnt=2
+            for i in range(int(folders*l),int((folders*l)+folders)):
+                floc=EXDIR+'/node_'+str(i+1)+'/data/zombie_1'
+                for j in range(2,inputs.zombs['ndet']+1):
+                    shutil.copy2(floc+"{:03d}".format(j)+".csv",EXDIR+'/node_'+str(l+1)+'_1/data/zombie_1'+"{:03d}".format(cnt)+".csv")
+                    cnt+=1
+    else:
+        # Step 2
+        shutil.copy2(EXDIR+'/node_1_1/data/zombie_1001.csv',EXDIR+'/data')
+        cnt=2
+        for l in range(subnodes):
+            for j in range(2,int((folders*(inputs.zombs['ndet']-1))+2)):
+                shutil.copy2(EXDIR+'/node_'+str(l+1)+'_1/data/zombie_1'+"{:03d}".format(j)+".csv",EXDIR+'/data/zombie_1'+"{:03d}".format(cnt)+".csv")
+                cnt+=1
+else:
 # # Single combination 
-# shutil.copy2(EXDIR+'/node_1/data/zombie_1001.csv',EXDIR+'/data')
-# cnt=2
-# for i in range(multflg):
-#     floc=EXDIR+'/node_'+str(i+1)+'/data/zombie_1'
-#     for j in range(2,inputs.zombs['ndet']+1):
-#         shutil.copy2(floc+"{:03d}".format(j)+".csv",EXDIR+'/data/zombie_1'+"{:03d}".format(cnt)+".csv")
-#         cnt+=1
+    shutil.copy2(EXDIR+'/node_1/data/zombie_1001.csv',EXDIR+'/data')
+    cnt=2
+    for i in range(multflg):
+        floc=EXDIR+'/node_'+str(i+1)+'/data/zombie_1'
+        for j in range(2,inputs.zombs['ndet']+1):
+            shutil.copy2(floc+"{:03d}".format(j)+".csv",EXDIR+'/data/zombie_1'+"{:03d}".format(cnt)+".csv")
+            cnt+=1
 
 cnt=cnt-1
 
@@ -134,9 +139,7 @@ with open("rundata.csv",'w',newline='')as file:
     writer.writerow([zomstat,hamstat,inputs.run['imagprop'],inputs.run['beta'],inputs.run['timesteps'],inputs.run['clean'],inputs.run['gram'],inputs.run['gramnum'],inputs.run['grad'],'n'])
     writer.writerow([inputs.zombs['norb'],inputs.zombs['nel'],inputs.zombs['spin'],cnt,inputs.zombs['zomtyp'],inputs.zombs['rhf_1'],inputs.zombs['imagflg']])
     writer.writerow([inputs.run['hamfile'],inputs.run['ovrlfile'],inputs.run['cleanham']])
-
-
-# Step 3
+ 
 os.environ["OMP_CANCELLATION"]="TRUE" 
 if(HPCFLG==1):
     if(inputs.run['cores']!=1):
@@ -154,39 +157,21 @@ if(HPCFLG==1):
     # f.write('time ./d_check.exe')
     f.write('time ./ZOMBIE.exe')
     f.close()
-   
-for l in range(subnodes):
-    shutil.copy2(file1, EXDIR+'/node_'+str(l+1)+'_1')
-    shutil.copy2("rundata.csv", EXDIR+'/node_'+str(l+1)+'_1')
-    shutil.copy2("ZOMBIE.exe", EXDIR+'/node_'+str(l+1)+'_1')
-    os.chdir(EXDIR+'/node_'+str(l+1)+'_1')
-    command=['qsub',file1]
-    subprocess.call(command)
+    if(stp1flg==1):
+        subprocess.call(['qsub',file1])
+    else:
+        shutil.copy2(file1, EXDIR+'/node_'+str(l+1)+'_1')
+        shutil.copy2("rundata.csv", EXDIR+'/node_'+str(l+1)+'_1')
+        shutil.copy2("ZOMBIE.exe", EXDIR+'/node_'+str(l+1)+'_1')
+        os.chdir(EXDIR+'/node_'+str(l+1)+'_1')
+        command=['qsub',file1]
+        subprocess.call(command)
 
-   
-# os.environ["OMP_CANCELLATION"]="TRUE" 
-# if(HPCFLG==1):
-#     if(inputs.run['cores']!=1):
-#         os.environ["OMP_NUM_THREADS"]=str(inputs.run['cores'])
-#     number=random.randint(99999,1000000)
-#     file1="zombie"+str(number)+".sh"
-#     f=open(file1,"w")
-#     f.write("#$ -cwd -V \n")
-#     if(inputs.run['cores']!=1):
-#         f.write("#$ -pe smp "+str(inputs.run['cores'])+" \n") #Use shared memory parallel environemnt 
-#     f.write("#$ -l h_rt="+inputs.run['runtime']+"\n")
-#     f.write("#$ -l h_vmem=5G \n")
-#     f.write("export OMP_CANCELLATION=true \n")
-#     f.write("module add mkl \n")
-#     # f.write('time ./d_check.exe')
-#     f.write('time ./ZOMBIE.exe')
-#     f.close()
-#     subprocess.call(['qsub',file1])
-# else:
-#     print(os.getcwd())
-#     if(inputs.run['cores']!=1):
-#         os.environ["OMP_NUM_THREADS"]=str(inputs.run['cores'])
-#     subprocess.run(["./ZOMBIE.exe"])
+else:
+    print(os.getcwd())
+    if(inputs.run['cores']!=1):
+        os.environ["OMP_NUM_THREADS"]=str(inputs.run['cores'])
+    subprocess.run(["./ZOMBIE.exe"])
